perf: use confirmation-modal in delete-tweet-modal

+++ src/features/tweets/components/delete-tweet-modal.tsx
@@ -1,65 +1,42 @@
-/* eslint-disable jsx-a11y/click-events-have-key-events */
-/* eslint-disable jsx-a11y/no-static-element-interactions */
+import { ConfirmationModal } from "@/components/elements/modal";
 import { useDisableBodyScroll } from "@/hooks";
 
 import { deleteMedia } from "../api/delete-media";
 import { useDeleteTweet } from "../hooks/use-delete-tweet";
 import { ITweet } from "../types";
 
-import styles from "./styles/delete-tweet-modal.module.scss";
-
 export const DeleteTweetModal = ({
   tweet,
   setIsDeleteModalOpen,
-  setIsActionsModalOpen,
+  setIsMenuOpen,
 }: {
   tweet: ITweet;
   setIsDeleteModalOpen: (value: boolean) => void;
-  setIsActionsModalOpen: (value: boolean) => void;
+  setIsMenuOpen: (value: boolean) => void;
 }) => {
   useDisableBodyScroll();
 
   const mutation = useDeleteTweet();
 
   return (
-    <div
-      onClick={() => {
+    <ConfirmationModal
+      heading="Delete Tweet?"
+      paragraph="This can’t be undone and it will be removed from your profile, the timeline of any accounts that follow you, and from Twitter search results."
+      confirmButtonText="Delete"
+      confirmButtonClick={() => {
+        mutation.mutate({
+          tweetId: tweet?.id,
+        });
         setIsDeleteModalOpen(false);
-        setIsActionsModalOpen(true);
+        if (tweet?.media?.length)
+          deleteMedia(tweet?.media?.map((m) => m.media_path));
       }}
-      className={styles.container}
-    >
-      <div onClick={(e) => e.stopPropagation()} className={styles.modal}>
-        <h1>Delete tweet?</h1>
-        <p>
-          This can’t be undone and it will be removed from your profile, the
-          timeline of any accounts that follow you, and from Twitter search
-          results.
-        </p>
-
-        <button
-          onClick={() => {
-            mutation.mutate({
-              tweetId: tweet?.id,
-            });
-            setIsDeleteModalOpen(false);
-            if (tweet?.media?.length)
-              deleteMedia(tweet?.media?.map((m) => m.media_path));
-          }}
-          className={styles.delete}
-        >
-          Delete
-        </button>
-        <button
-          className={styles.cancel}
-          onClick={() => {
-            setIsDeleteModalOpen(false);
-            setIsActionsModalOpen(true);
-          }}
-        >
-          Cancel
-        </button>
-      </div>
-    </div>
+      confirmButtonStyle="delete"
+      cancelButtonText="Cancel"
+      cancelButtonClick={() => {
+        setIsDeleteModalOpen(false);
+        setIsMenuOpen(true);
+      }}
+    />
   );
 };
